Move function into test harness
